+++ services/index/internal/adapter/elasticsearch/client.go
