+++ services/index/internal/adapter/mongo/nodes_db/nodes_db.go
