+++ services/index/internal/adapter/nats/nats.go
