+++ services/index/internal/app/cleanup.go
@@ -1,6 +1,6 @@
 package app
 
-import "github.com/MurmurationsNetwork/MurmurationsServices/services/index/internal/datasources/mongo/nodes_db"
+import "github.com/MurmurationsNetwork/MurmurationsServices/services/index/internal/adapter/mongo/nodes_db"
 
 func cleanup() {
 	nodes_db.Disconnect()
