+++ services/validation/internal/adapter/nats/nats.go
