+++ services/validation/internal/app/cleanup.go
@@ -1,6 +1,6 @@
 package app
 
-import "github.com/MurmurationsNetwork/MurmurationsServices/services/validation/internal/datasource/nats"
+import "github.com/MurmurationsNetwork/MurmurationsServices/services/validation/internal/adapter/nats"
 
 func cleanup() {
 	nats.Disconnect()
